+++ NEWS.md
@@ -1,4 +1,4 @@
-# cookies (development version)
+# cookies 0.2.2
 
 * Updated cookie handling to allow HttpOnly cookies mixed with JavaScript cookies, and improved error messages (#54).
 * Updated cookie handling to find cookies even if the call comes from inside a module (#56).